requirements.txt
# Python Dependencies for Bibliometric Analysis System

# Web Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# LLM APIs (Required for classification)
openai>=1.40.0
anthropic>=0.18.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Web Crawling
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
scholarly==1.7.11

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1

# Database (install separately when needed - requires PostgreSQL)
# psycopg2-binary==2.9.9
# pymongo==4.6.1
# sqlalchemy==2.0.25

# Testing (optional)
# pytest==7.4.4
# pytest-cov==4.1.0
